refactor: extract Resolution into mod
